requirements.txt
opencv-python>=4.5.1
numpy>=1.19.5
torch>=1.7.1
torchvision>=0.8.2
matplotlib>=3.3.4
Pillow>=8.1.0
scikit-learn>=0.24.1
scipy>=1.6.0
pyexifinfo>=0.2.1
ultralytics>=0.6.5